[refactor]: Added documentation and moved classes
